refactor: move stitches to separated folder
