+++ src/lib/stitches/index.ts
@@ -1,0 +1,1 @@
+export * from './stitches-config'
